// apps/curator-web/src/api/generated/schema.d.ts
export type components = {
    schemas: {
// ...
        CreateUploadSessionRequest: {
            /**
             * Access Scope Id
             * Format: uuid
             */
            access_scope_id: string;
            /** Byte Size */
            byte_size: number;
            /** Document Id */
            document_id?: string | null;
            /** Document Number */
            document_number?: string | null;
            /**
             * License Policy Id
             * Format: uuid
             */
            license_policy_id: string;
            /**
             * Media Type
             * @default application/pdf
             */
            media_type: string;
            /** Original Filename */
            original_filename: string;
            /**
             * Project Id
             * Format: uuid
             */
            project_id: string;
            /** Revision Label */
            revision_label: string;
            /**
             * Source Organization Id
             * Format: uuid
             */
            source_organization_id: string;
            /** Title */
            title?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
};